+++ Assets/Scripts/Mobs/Stats/MobStats.cs
@@ -2,14 +2,14 @@ using UnityEngine;
 
 public class MobStats : MonoBehaviour
 {
-    public short maxHealth;
-    public float currentHealth;
-    public float healthRegenTime;
+    public int maxHealth;
+    public int currentHealth;
+    public int healthRegenTime;
 
     public float baseMovementSpeed;
     public float sprintMovementSpeed;
 
-    public ushort damage;
+    public int damage;
     public string damageType;
 
     public float attackRate;
